refactor(web): extract admin audit route shell
Move admin audit page logic into a feature module so the route shell stays within the strict architecture budget enforced by CI.

Refs #325
Refs #329
Refs #336
(cherry picked from commit dfbd69e4487e64b9bcb5715b88a6a7928531817c)
